[llm, entity resolution]: Update repo structure
