+++ kif_llm_store/store/llm/entity_resolution/__init__.py
@@ -1,11 +1,12 @@
 # Copyright (C) 2024 IBM Corp.
 # SPDX-License-Identifier: Apache-2.0
 
-from .abc import Disambiguator, EntitySource
-from .wikidata_entity_source import WikidataEntitySource
-from .dbpedia_entity_source import DBpediaEntitySource
-from .naive_disambiguator import NaiveDisambiguator
-from .llm_disambiguator import LLM_Disambiguator
+from .disambiguators.abc import Disambiguator
+from .entity_sources.abc import EntitySource
+from .entity_sources.wikidata_entity_source import WikidataEntitySource
+from .entity_sources.dbpedia_entity_source import DBpediaEntitySource
+from .disambiguators.naive_disambiguator import NaiveDisambiguator
+from .disambiguators.llm_disambiguator import LLM_Disambiguator
 
 __all__ = (
     'Disambiguator',
