+++ kif_llm_store/store/llm/entity_resolution/disambiguators/similarity_disambiguator.py
@@ -9,14 +9,13 @@ from kif_lib import Item, Property
 from kif_lib.typing import Any, Callable, Iterator, Optional
 from kif_lib.vocabulary import wd
 
-from constants import (
+from ..constants import (
     PID,
     QID,
     WID,
     Label,
 )
 from .abc import Disambiguator
-from .util import fetch_wikidata_entities
 
 LOG = logging.getLogger(__name__)
