+++ kif_llm_store/store/llm/entity_resolution/entity_sources/wikidata_entity_source.py
@@ -13,85 +13,89 @@ from typing import (
     Optional,
 )
 
-from ..constants import (
-    DBPEDIA_SEARCH_API_BASE_URL,
+from ...constants import (
+    WIKIDATA_SEARCH_API_BASE_URL,
 )
 
-from kif_lib.namespace.dbpedia import DBpedia
+from kif_lib.namespace.wikidata import Wikidata
 
 LOG = logging.getLogger(__name__)
 
 nest_asyncio.apply()
 
 
-class DBpediaEntitySource(
+class WikidataEntitySource(
     EntitySource,
-    source_name='dbpedia',
+    source_name='wikidata',
 ):
 
-    #: The default DBpedia API endpoint IRI.
-    _default_api_uri: str = DBPEDIA_SEARCH_API_BASE_URL
+    #: The default Wikidata API endpoint IRI.
+    _default_api_uri: str = WIKIDATA_SEARCH_API_BASE_URL
 
     _api_uri: Optional[str]
 
-    _default_prefix_item_iri = str(DBpedia.RESOURCE)
-    _default_prefix_property_iri = str(DBpedia.PROPERTY)
+    _default_prefix_item_iri = str(Wikidata.WD)
+    _default_prefix_property_iri = str(Wikidata.WDT)
 
     def __init__(
         self, source_name: str, uri: Optional[str] = None, **kwargs: Any
     ) -> None:
         assert source_name == self.source_name
 
+        self.default_prefix_item_iri = self._default_prefix_item_iri
+        self.default_prefix_property_iri = self._default_prefix_property_iri
+
         self._api_uri = uri or self._default_api_uri
         super().__init__(source_name, **kwargs)
 
-    async def _get_items_from_label(
+    def _lookup_item_search(
         self, label: str, limit=10
     ) -> dict[str, Optional[Any]]:
-        return await self.__get_entity_from_label(label, 'item', limit)
+        return self.__get_entity_from_label(label, 'item', limit)
 
-    async def _get_properties_from_label(
+    def _lookup_property_search(
         self, label, limit
     ) -> dict[str, Optional[Any]]:
-        return await self.__get_entity_from_label(label, 'property', limit)
+        return self.__get_entity_from_label(label, 'property', limit)
 
     async def __get_entity_from_label(
         self, label: str, type: str, limit: int
     ) -> dict[str, Optional[Any]]:
         import httpx
-        import re
 
         assert type in ['item', 'property']
 
         template = (
             f'{self._api_uri}'
-            '?query={label}'
-            '&format=JSON'
-            '&maxResults={limit}'
+            '?action=wbsearchentities'
+            '&search={label}'
+            '&language=en'
+            '&format=json'
+            '&limit={limit}'
+            '&type={type}'
         )
 
-        url = template.format_map(defaultdict(str, label=label, limit=limit))
+        url = template.format_map(
+            defaultdict(str, label=label, limit=limit, type=type)
+        )
 
         try:
             async with httpx.AsyncClient() as client:
                 response = await client.get(url)
                 data = response.json()
-                search_result = data.get('docs', [])
-                a = [
+                search_result = data.get('search', [])
+                return [
                     {
-                        'id': result.get('id', [''])[0],
-                        'iri': result.get('resource', [''])[0],
-                        'label': re.sub(
-                            r"<.*?>", "", result.get('label', [''])[0]
+                        'id': result['id'],
+                        'iri': result.get(
+                            'concepturi',
+                            f"http://www.wikidata.org/entity/{result['id']}",
                         ),
-                        'description': re.sub(
-                            r"<.*?>", "", result.get('comment', [''])[0]
-                        ),
+                        'label': result.get('label', ''),
+                        'description': result.get('description', ''),
                     }
                     for result in search_result
                 ]
-
-                return a
         except httpx.RequestError as e:
             LOG.error('Request error for label "%s": %s', label, e)
             raise e
@@ -106,3 +110,15 @@ class DBpediaEntitySource(
         except Exception as e:
             LOG.error(f'Unexpected error for label `{label}`: {e}')
             raise e
+
+    def _parse_entity(self, id: str, entity_type='item') -> Optional[str]:
+        import re
+
+        fl = 'Q'
+        if entity_type == 'property':
+            fl = 'P'
+        match = re.search(fr'{fl}\d+', id)
+
+        if match:
+            return match.group()
+        return None
